rag/card_recommend/requirements.txt
fastapi
uvicorn
pydantic
weaviate-client
streamlit